update segment, and add tests
update segment, and add tests
