+++ plugins/Ecommerce/tests/System/expected/test_ecommerceOrderWithItems_SegmentCartRevenueOrder__Goals.get_day.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<result>
+    <nb_conversions>2</nb_conversions>
+    <nb_visits_converted>2</nb_visits_converted>
+    <revenue>5020.22</revenue>
+    <items>8</items>
+    <conversion_rate>100%</conversion_rate>
+    <nb_conversions_new_visit>0</nb_conversions_new_visit>
+    <nb_visits_converted_new_visit>0</nb_visits_converted_new_visit>
+    <revenue_new_visit>0</revenue_new_visit>
+    <items_new_visit>0</items_new_visit>
+    <conversion_rate_new_visit>0%</conversion_rate_new_visit>
+    <nb_conversions_returning_visit>2</nb_conversions_returning_visit>
+    <nb_visits_converted_returning_visit>2</nb_visits_converted_returning_visit>
+    <revenue_returning_visit>5020.22</revenue_returning_visit>
+    <items_returning_visit>8</items_returning_visit>
+    <conversion_rate_returning_visit>100%</conversion_rate_returning_visit>
+</result>
