+++ plugins/Ecommerce/tests/System/expected/test_ecommerceOrderWithItems_SegmentRevenueOrder__Goals.get_day.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0"?>
+<result>
+    <nb_conversions>2</nb_conversions>
+    <nb_visits_converted>1</nb_visits_converted>
+    <revenue>3111.11</revenue>
+    <conversion_rate>50%</conversion_rate>
+    <nb_conversions_new_visit>0</nb_conversions_new_visit>
+    <nb_visits_converted_new_visit>0</nb_visits_converted_new_visit>
+    <revenue_new_visit>0</revenue_new_visit>
+    <conversion_rate_new_visit>0%</conversion_rate_new_visit>
+    <nb_conversions_returning_visit>2</nb_conversions_returning_visit>
+    <nb_visits_converted_returning_visit>1</nb_visits_converted_returning_visit>
+    <revenue_returning_visit>3111.11</revenue_returning_visit>
+    <conversion_rate_returning_visit>50%</conversion_rate_returning_visit>
+</result>
